+++ internal/xsql/common/rows.go
@@ -1,0 +1,9 @@
+package common
+
+import "database/sql/driver"
+
+type Rows interface {
+	driver.RowsNextResultSet
+	driver.RowsColumnTypeDatabaseTypeName
+	driver.RowsColumnTypeNullable
+}