+++ internal/xsql/xtable/rows.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/scanner"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/common"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/result"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/result/indexed"
@@ -18,7 +18,7 @@ import (
 
 var (
 	_ driver.Rows                           = &rows{}
-	_ driver.RowsNextResultSet              = &rows{}
+	_ common.Rows                           = &rows{}
 	_ driver.RowsColumnTypeDatabaseTypeName = &rows{}
 	_ driver.RowsColumnTypeNullable         = &rows{}
 
@@ -95,7 +95,7 @@ func (r *rows) NextResultSet() (finalErr error) {
 	r.nextSet.Do(func() {})
 	err := r.result.NextResultSetErr(context.Background())
 	if err != nil {
-		return badconn.Map(xerrors.WithStackTrace(err))
+		return xerrors.WithStackTrace(err)
 	}
 
 	return nil
@@ -111,10 +111,10 @@ func (r *rows) Next(dst []driver.Value) error {
 		err = r.result.NextResultSetErr(context.Background())
 	})
 	if err != nil {
-		return badconn.Map(xerrors.WithStackTrace(err))
+		return xerrors.WithStackTrace(err)
 	}
 	if err = r.result.Err(); err != nil {
-		return badconn.Map(xerrors.WithStackTrace(err))
+		return xerrors.WithStackTrace(err)
 	}
 	if !r.result.NextRow() {
 		return io.EOF
@@ -124,7 +124,7 @@ func (r *rows) Next(dst []driver.Value) error {
 		values[i] = &valuer{}
 	}
 	if err = r.result.Scan(values...); err != nil {
-		return badconn.Map(xerrors.WithStackTrace(err))
+		return xerrors.WithStackTrace(err)
 	}
 	for i := range values {
 		val, ok := values[i].(*valuer)
@@ -135,7 +135,7 @@ func (r *rows) Next(dst []driver.Value) error {
 		dst[i] = val.Value()
 	}
 	if err = r.result.Err(); err != nil {
-		return badconn.Map(xerrors.WithStackTrace(err))
+		return xerrors.WithStackTrace(err)
 	}
 
 	return nil